Sets up the neighbourhood of each pixel on the grid.
The neighbourhood is a 3x3 block (2d-vector), with pointers to the neighbours. In the case of an edge pixel (ex. the one at position 0,0) "the pointers to pixels" that fall out of the grid are nullptr.
Testing code is required, but there is a small demonstration that will be wiped out in future commits.